backend/pyproject.toml
[tool.poetry]
name = "simple-upload"
version = "0.1.0"
description = ""
authors = ["zhangjian <zhangjian@zingfront.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
loguru = "^0.7.2"
pydantic-settings = "^2.6.1"
redis = "^5.2.0"
asgiref = "^3.8.1"
msgspec = "^0.18.6"
cryptography = "^43.0.3"
itsdangerous = "^2.2.0"
sqlalchemy-crud-plus = "^1.6.0"
pydantic-core = "^2.26.0"
mysqlclient = "^2.2.5"
fastapi-limiter = "^0.1.6"
fastapi-pagination = "^0.12.31"
pydantic-extra-types = "^2.10.0"
phonenumbers = "^8.13.49"
asyncmy = "^0.2.9"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
